services/platform-api/requirements.txt
# services/platform-api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
httpx>=0.26
supabase>=2.0
docling>=2.70.0
eyecite>=2.7
javalang>=0.13.0
tree-sitter>=0.24
tree-sitter-java>=0.23
tree-sitter-python>=0.23
tree-sitter-javascript>=0.23
tree-sitter-typescript>=0.23
tree-sitter-go>=0.23
tree-sitter-rust>=0.23
tree-sitter-c-sharp>=0.23
pymongo>=4.6,<5
lxml>=5.0
PyJWT>=2.8
cryptography>=42.0.0
msgspec>=0.18
more-itertools>=10.0
Jinja2>=3.1
pytest>=8.0
pytest-asyncio>=0.23
google-cloud-storage>=2.19
opentelemetry-api>=1.28
opentelemetry-sdk>=1.28
opentelemetry-exporter-otlp-proto-http>=1.28
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-instrumentation-httpx>=0.49b0
opentelemetry-instrumentation-logging>=0.49b0
mistune>=3.0
numpy>=1.26
